Refactor: move create tag functionality out of updateRespositoryPointerToBranch
